collections/ansible_collections/lv3/platform/roles/mail_platform_runtime/files/mail-gateway/requirements.txt
fastapi==0.118.2
httpx==0.28.1
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-instrumentation-httpx==0.60b0
uvicorn==0.37.0
sentry-sdk[fastapi]==2.42.1
